requirements.txt
easygui==0.98.3
numpy==2.2.5
pillow==11.2.1
